+++ src/main/java/pl/ergohestia/ehj1/ivesta/controller/RouteController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.ergohestia.ehj1.ivesta.controller.request.RouteRequest;
+import pl.ergohestia.ehj1.ivesta.request.RouteRequest;
 import pl.ergohestia.ehj1.ivesta.model.RouteDto;
 import pl.ergohestia.ehj1.ivesta.request.DriverAssociation;
 import pl.ergohestia.ehj1.ivesta.request.VehicleAssociation;
